AzTFServerless/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
bleach==1.5.0
boto3==1.4.7
botocore==1.7.48
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
enum34==1.1.6
grpcio==1.14.2
grpcio-tools==1.14.2
html5lib==0.9999999
idna==2.8
jmespath==0.9.3
Markdown==3.0.1
numpy==1.16.0
protobuf==3.6.1
python-dateutil==2.7.5
requests==2.21.0
requests-toolbelt==0.8.0
s3transfer==0.1.13
six==1.12.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
urllib3==1.24.1
Werkzeug==0.14.1
